Simplify.
These values come from SFDC. To simplify querying, use the value as the
ID.  The exception is, if there is a slash, it must be converted
because Nuxeo does not allow slashes in vocal ids.  But this is simpler
than creating a full mapping between SFDC and Nuxeo on the client.